Stage 1: Unified Build
