+++ Sources/Bedrockifier/Commands/Root.swift
@@ -1,7 +1,7 @@
 /*
  Bedrockifier
 
- Copyright (c) 2021 Adam Thayer
+ Copyright (c) 2026 Adam Thayer
  Licensed under the MIT license, as follows:
 
  Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -23,27 +23,13 @@
  SOFTWARE.)
  */
 
-import ConsoleKit
-import Foundation
+import ArgumentParser
 import Logging
 
-import Bedrockifier
-
-// Configure Console
-let terminal = Terminal()
-var input = CommandInput(arguments: CommandLine.arguments)
-
-var commands = Commands()
-commands.use(PackCommand(), as: "pack")
-commands.use(TrimCommand(), as: "trim")
-commands.use(UnpackCommand(), as: "unpack")
-var allCommands = commands.group(help: "Minecraft Bedrock Backup Tool")
-
-LoggingSystem.bootstrap(console: terminal, level: .trace, metadata: .init())
-
-do {
-    try terminal.run(allCommands, input: input)
-} catch let error {
-    terminal.error(error.localizedDescription)
-    exit(1)
+@main
+struct Bedrockifier: AsyncParsableCommand {
+    static let configuration = CommandConfiguration(
+        abstract: "A utility for backing up Minecraft servers.",
+        subcommands: [Service.self, Pack.self, Trim.self, Unpack.self],
+    )
 }
