+++ Sources/Bedrockifier/Service/BackupActor.swift
@@ -24,7 +24,6 @@
  */
 
 import Foundation
-import Bedrockifier
 
 actor BackupActor {
     let config: BackupConfig
